# 🦟 Bug fix

Addresses #429

## Summary
<!-- Describe your fix, including an explanation of how to reproduce the bug
before and after the PR.-->

The test rarely fails, but when it does, it looks like it's just during the 1st iteration, where there's only one matching entity and one non-matching, with a difference of a few nanoseconds:

```
147: [ RUN      ] EntityComponentManagerPerfrormance.Each
147: /var/lib/jenkins/workspace/ignition_gazebo-ci-ign-gazebo3-bionic-amd64/ign-gazebo/test/performance/each.cc:128: Failure
147: Expected: (cacheEntityAvg) < (cachelessEntityAvg), actual: 417.82 vs 386.29
147: Matching Entity Count =		1
147: Nonmatching Entity Count =	1
147: Each Iterations =		100
147: Cache total =			41782 ns
147: Cache avg per iter =		417.81999999999999 ns
147: Cache avg per iter*entity =	417.81999999999999 ns
147: Cacheless total =		38629 ns
147: Cacheless avg per iter=		386.29000000000002 ns
147: Cacheless avg per iter*entity=	386.29000000000002 ns
147: 
147: [  FAILED  ] EntityComponentManagerPerfrormance.Each (3329 ms)
```

I think that this isn't unexpected, the performance advantages of the cached `Each` call are most noticeable with more entities. The lower the number of entities, the more the test may be affected by random variation on the CI machine at the time the test is running. 

So I think it's reasonable to relax the test a bit.

## Checklist
- [x] Signed all commits for DCO
- [x] Added tests
- [ ] ~~Updated documentation (as needed)~~
- [ ] ~~Updated migration guide (as needed)~~
- [x] `codecheck` passed (See [contributing](https://ignitionrobotics.org/docs/all/contributing#contributing-code))
- [x] All tests passed (See [test coverage](https://ignitionrobotics.org/docs/all/contributing#test-coverage))
- [x] While waiting for a review on your PR, please help review [another open pull request](https://github.com/pulls?q=is%3Aopen+is%3Apr+user%3Aignitionrobotics+repo%3Aosrf%2Fsdformat+archived%3Afalse+) to support the maintainers

**Note to maintainers**: Remember to use **Squash-Merge**

---

https://github.com/osrf/buildfarmer/issues/156